refactor : adminId를 Long으로 받아오도록 수정
