+++ src/main/java/com/example/project/solution/dto/request/admin/SolutionDeleteRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class SolutionDeleteRequest {
-    private String adminId;
+    private Long adminId;
     private Long solutionId;
 }
