+++ src/main/java/com/example/project/solution/dto/request/admin/SolutionRegisterRequest.java
@@ -7,7 +7,7 @@ import lombok.Data;
 
 @Data
 public class SolutionRegisterRequest implements RegisterRequest<Solution> {
-    private String adminId;
+    private Long adminId;
     private Difficulty difficulty;
     private String title;
     private String description;
